Implemented a simple command line argument parser that is backwards compatible with the simple command line options specified from before.
Options may now be specified using the double dash prefix, i.e.
  --pre=scripts/my_pre_init_script.mac
to specify the pre-init script macro. Unknown options are simply
ignored, for now. But if no double dash parameters are specified,
we revert to the old way of parsing the command line arguments.
Which is, first parameter is the preinit macro, and second is the
post init macro.

Three command line options exist right now:
--pre=<script_name>
--post=<script_name>
--gui (no arguments)
The --pre and --post work as expected, and the --gui is a flag that
forces the use of the GUI even if pre and post scripts are specified.

The reason for these changes is to allow one to specify pre and post
scripts on the command line while simultaniously using the GUI.
This is helpful while debugging the geometry of detectors, for example.

Other options can be implemented as necessary.