Fix "shadow" warnings.
The most annoying warning was with the seconds 's' in G4SystemOfUnits and
an internal 's' in a TString.  The quick fix was to move the inclusion of
G4SystemOfUnits to the end in the source files, and to remove it
entirely from G4SBSIO (relying now on using the full namespace of CLHEP, i.e.,
CLHEP::GeV).