Modified the way to build the Grinch tank: It use to be a simple box with entrance and exit window. Now, it is a box with a wedge cut on the corner where the PMTs are located, extended with a box to contain the PMT box. The extension for the PMT box is necessary, as it is much simpler to include the PMTs, which are sensitive detectors, "inside" the gas volume. Also, the tank gas is now constructed similarly, with the notable difference that it is full instead of being empty.
In the details: we build both the "outer box" and the "inner box"
in the way described before.
If we construct the tank box, the inner box is subtracted from the outer box
(with G4SubtractionSolid), then the outer box is returned.
If we construct the gas volume, then, the inner box only is returned.

WARNING: This still needs some fine tuning,
and ought to be optimised before being merged.