Major commit of new g4sbs developments with "track information" machinery:
I. Includes new "user track information", "user tracking action", and "user stepping action" classes with the primary goal being to associate relevant particle track information with "hits" in sensitive detectors. The three major types of information to be recorded include:

1. "Primary" track information: this associates any given sensitive detector hit with the generated primary particle at the interaction vertex that ultimately caused it, and records basic information (PID, position, momentum, polarization, G4 track ID and parent ID) about said track in an efficient branch structure in the ROOT tree.
2. "Original" track information: this associates any given sensitive detector hit with either the primary particle or the secondary particle responsible for causing that hit, where if the particle is a secondary, it is recorded if and only if it is produced within a logical volume that has been flagged as either a "target" or "analyzer" volume (or both). The typical use case for this information would be for background and radiation shielding studies, and/or for detailed analysis of recoil nucleon polarimetry in experiments of that type. The properties of the particle at its production vertex are recorded.
3. "Sensitive detector boundary crossing" track information. This associates any given sensitive detector hit with basic information about the particle track as it enters a (hopefully sensibly defined) "boundary volume" which is typically a mother logical volume enclosing (and containing) a sensitive detector logical volume. The association between sensitive detector logical volumes and boundary volumes is defined by the user when the geometry of the detector is coded. For example, the entire HCAL is enclosed within a box that contains the entire detector, and in the code that constructs the HCAL geometry, this "mother box" is added to an STL map where the key is the sensitive detector name and the mapped value is the logical volume name of the "boundary" volume. The main use case for this information is to evaluate detector performance in terms of energy and coordinate resolution by comparing the properties of particles as they enter the detectors to the properties reconstructed from the signals recorded by the detectors.

This information is stored in the ROOT tree in dedicated array-valued branches organized by sensitive detector name such that the information for each unique track of type 1, 2, or 3 appears exactly once for each sensitive detector with hits which that track is ultimately involved in creating. Then, an integer index referring to a position in the arrays of type 1, 2, and 3 track information is stored within the "hit" arrays for each sensitive detector so that the correct track information can be referred to for each hit individually. Typically, only one or a few tracks of each type will be involved per event in any given sensitive detector, and it is often, and perhaps usually the case that the same track will be involved in causing many hits in several sensitive detectors, so there is some redundancy of information.

The new user command "/g4sbs/keepsdtrackinfo were added to enable/disable the recording of this information, and it is DISABLED by default. It can be enabled individually by sensitive detector name, or enabled for all detectors by using "all" in place of individual detector names.

II. Convenience features:

1. Modified G4SBSPhysicsList such that optical photon production is DISABLED by default, and has to be turned ON by the user using the /g4sbs/usescint and /g4sbs/useckov commands in the "pre-initialization" macro.

III. New utility macro added for "quick and dirty" figure-of-merit estimation for the GEP experiment. Does basic event rate and polarimetry figure-of-merit estimate using a parametrization of p+CH2 analyzing power and the proton form factors from Puckett et al., PRC 2017